libdnf5 and libdnf5-cli: Move exception declarations to own header files
Some of the exceptions have already been declared in their own header
files. But not all of them.  This PR unifies the approach. Separating
exception declarations from other declarations simplifies their
subsequent use, such as wrapping in SWIG.
