+++ include/libdnf5/base/transaction_errors.hpp
@@ -17,24 +17,26 @@ You should have received a copy of the GNU Lesser General Public License
 along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#ifndef LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
-#define LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
 
-#include "transaction_item_errors.hpp"
+#ifndef LIBDNF5_BASE_TRANSACTION_ERRORS_HPP
+#define LIBDNF5_BASE_TRANSACTION_ERRORS_HPP
 
+#include "libdnf5/common/exception.hpp"
 #include "libdnf5/defs.h"
 
-#include <string>
 
+namespace libdnf5::base {
 
-namespace libdnf5::transaction {
+/// Error related to processing transaction
+class LIBDNF_API TransactionError : public Error {
+public:
+    using Error::Error;
+    /// @return Error class' domain name"
+    const char * get_domain_name() const noexcept override { return "libdnf5::base"; }
+    /// @return Error class' name"
+    const char * get_name() const noexcept override { return "TransactionError"; }
+};
 
-enum class TransactionItemState : int { STARTED = 1, OK = 2, ERROR = 3 };
+}  // namespace libdnf5::base
 
-
-LIBDNF_API std::string transaction_item_state_to_string(TransactionItemState state);
-LIBDNF_API TransactionItemState transaction_item_state_from_string(const std::string & state);
-
-}  // namespace libdnf5::transaction
-
-#endif  // LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
+#endif  // LIBDNF5_BASE_TRANSACTION_ERRORS_HPP
