+++ include/libdnf5/conf/option.hpp
@@ -20,7 +20,8 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #ifndef LIBDNF5_CONF_OPTION_HPP
 #define LIBDNF5_CONF_OPTION_HPP
 
-#include "libdnf5/common/exception.hpp"
+#include "option_errors.hpp"
+
 #include "libdnf5/common/impl_ptr.hpp"
 #include "libdnf5/defs.h"
 
@@ -28,36 +29,6 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 
 
 namespace libdnf5 {
-
-/// Option exception
-class LIBDNF_API OptionError : public Error {
-public:
-    using Error::Error;
-    const char * get_domain_name() const noexcept override { return "libdnf5"; }
-    const char * get_name() const noexcept override { return "OptionError"; }
-};
-
-/// Exception that is generated when an invalid input value is detected.
-class LIBDNF_API OptionInvalidValueError : public OptionError {
-public:
-    using OptionError::OptionError;
-    const char * get_name() const noexcept override { return "OptionInvalidValueError"; }
-};
-
-/// Exception that is generated when not allowed input value is detected.
-class LIBDNF_API OptionValueNotAllowedError : public OptionInvalidValueError {
-public:
-    using OptionInvalidValueError::OptionInvalidValueError;
-    const char * get_name() const noexcept override { return "OptionValueNotAllowedError"; }
-};
-
-/// Exception that is generated during read an empty Option.
-class LIBDNF_API OptionValueNotSetError : public OptionError {
-public:
-    using OptionError::OptionError;
-    const char * get_name() const noexcept override { return "OptionValueNotSetError"; }
-};
-
 
 /// Option class is an abstract class. Parent of all options. Options are used to store a configuration.
 // @replaces libdnf:conf/Option.hpp:class:Option
