+++ include/libdnf5/conf/option_path_errors.hpp
@@ -17,24 +17,22 @@ You should have received a copy of the GNU Lesser General Public License
 along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#ifndef LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
-#define LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
+#ifndef LIBDNF5_CONF_OPTION_PATH_ERRORS_HPP
+#define LIBDNF5_CONF_OPTION_PATH_ERRORS_HPP
 
-#include "transaction_item_errors.hpp"
+#include "option_errors.hpp"
 
 #include "libdnf5/defs.h"
 
-#include <string>
+namespace libdnf5 {
 
+/// Exception that is generated when input path does not exist.
+class LIBDNF_API OptionPathNotFoundError : public OptionValueNotAllowedError {
+public:
+    using OptionValueNotAllowedError::OptionValueNotAllowedError;
+    const char * get_name() const noexcept override { return "OptionPathNotFoundError"; }
+};
 
-namespace libdnf5::transaction {
+}  // namespace libdnf5
 
-enum class TransactionItemState : int { STARTED = 1, OK = 2, ERROR = 3 };
-
-
-LIBDNF_API std::string transaction_item_state_to_string(TransactionItemState state);
-LIBDNF_API TransactionItemState transaction_item_state_from_string(const std::string & state);
-
-}  // namespace libdnf5::transaction
-
-#endif  // LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
+#endif
