+++ include/libdnf5/conf/vars_errors.hpp
@@ -17,24 +17,23 @@ You should have received a copy of the GNU Lesser General Public License
 along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#ifndef LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
-#define LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
+#ifndef LIBDNF5_CONF_VARS_ERRORS_HPP
+#define LIBDNF5_CONF_VARS_ERRORS_HPP
 
-#include "transaction_item_errors.hpp"
-
+#include "libdnf5/common/exception.hpp"
 #include "libdnf5/defs.h"
 
-#include <string>
 
+namespace libdnf5 {
 
-namespace libdnf5::transaction {
+// Thrown when attempting to set a read-only variable
+class LIBDNF_API ReadOnlyVariableError : public Error {
+    using Error::Error;
 
-enum class TransactionItemState : int { STARTED = 1, OK = 2, ERROR = 3 };
+    const char * get_domain_name() const noexcept override { return "libdnf5"; }
+    const char * get_name() const noexcept override { return "ReadOnlyVariableError"; }
+};
 
+}  // namespace libdnf5
 
-LIBDNF_API std::string transaction_item_state_to_string(TransactionItemState state);
-LIBDNF_API TransactionItemState transaction_item_state_from_string(const std::string & state);
-
-}  // namespace libdnf5::transaction
-
-#endif  // LIBDNF5_TRANSACTION_TRANSACTION_ITEM_STATE_HPP
+#endif
