+++ include/libdnf5/repo/repo_errors.hpp
@@ -21,51 +21,52 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #define LIBDNF5_RPM_REPO_ERRORS_HPP
 
 #include "libdnf5/common/exception.hpp"
+#include "libdnf5/defs.h"
 
 
 namespace libdnf5::repo {
 
-class RepoError : public Error {
+class LIBDNF_API RepoError : public Error {
     using Error::Error;
     const char * get_domain_name() const noexcept override { return "libdnf5::repo"; }
     const char * get_name() const noexcept override { return "RepoError"; }
 };
 
 
-class RepoCacheonlyError : public RepoError {
+class LIBDNF_API RepoCacheonlyError : public RepoError {
 public:
     using RepoError::RepoError;
     const char * get_name() const noexcept override { return "RepoCacheonlyError"; }
 };
 
 
-class RepoDownloadError : public RepoError {
+class LIBDNF_API RepoDownloadError : public RepoError {
 public:
     using RepoError::RepoError;
     const char * get_name() const noexcept override { return "RepoDownloadError"; }
 };
 
 
-class RepoPgpError : public RepoError {
+class LIBDNF_API RepoPgpError : public RepoError {
 public:
     using RepoError::RepoError;
     const char * get_name() const noexcept override { return "RepoPgpError"; }
 };
 
 
-class RepoRpmError : public RepoError {
+class LIBDNF_API RepoRpmError : public RepoError {
     using RepoError::RepoError;
     const char * get_name() const noexcept override { return "RepoRpmError"; }
 };
 
 
-class RepoCompsError : public RepoError {
+class LIBDNF_API RepoCompsError : public RepoError {
     using RepoError::RepoError;
     const char * get_name() const noexcept override { return "RepoCompsError"; }
 };
 
 
-class RepoIdAlreadyExistsError : public RepoError {
+class LIBDNF_API RepoIdAlreadyExistsError : public RepoError {
     using RepoError::RepoError;
     const char * get_name() const noexcept override { return "RepoIdAlreadyExistsError"; }
 };
