+++ include/libdnf5/transaction/transaction.hpp
@@ -23,6 +23,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #include "comps_environment.hpp"
 #include "comps_group.hpp"
 #include "rpm_package.hpp"
+#include "transaction_errors.hpp"
 #include "transaction_item.hpp"
 
 #include "libdnf5/base/transaction_environment.hpp"
@@ -60,14 +61,6 @@ enum class TransactionState : int { STARTED = 1, OK = 2, ERROR = 3 };
 
 LIBDNF_API std::string transaction_state_to_string(TransactionState state);
 LIBDNF_API TransactionState transaction_state_from_string(const std::string & state);
-
-class LIBDNF_API InvalidTransactionState : public libdnf5::Error {
-public:
-    InvalidTransactionState(const std::string & state);
-
-    const char * get_domain_name() const noexcept override { return "libdnf5::transaction"; }
-    const char * get_name() const noexcept override { return "InvalidTransactionState"; }
-};
 
 
 /// Transaction holds information about a transaction.
